Removed const file - Moved constants to their respective files -
