+++ tibiapy/__init__.py
@@ -1,6 +1,5 @@
 from tibiapy import abc, utils
 from tibiapy.character import Character, Death, Killer, OnlineCharacter, OtherCharacter
-from tibiapy.const import *
 from tibiapy.errors import *
 from tibiapy.guild import Guild, GuildInvite, GuildMember
 from tibiapy.world import World, WorldOverview
