+++ tibiapy/const.py
@@ -1,8 +1,3 @@
-CHARACTER_URL = "https://www.tibia.com/community/?subtopic=characters&name="
-CHARACTER_URL_TIBIADATA = "https://api.tibiadata.com/v2/characters/%s.json"
-GUILD_URL = "https://www.tibia.com/community/?subtopic=guilds&page=view&GuildName="
-GUILD_URL_TIBIADATA = "https://api.tibiadata.com/v2/guild/%s.json"
-GUILD_LIST_URL = "https://www.tibia.com/community/?subtopic=guilds&world="
-GUILD_LIST_URL_TIBIADATA = "https://api.tibiadata.com/v2/guilds/%s.json"
-WORLD_URL = "https://www.tibia.com/community/?subtopic=worlds&world=%s"
-WORLD_URL_TIBIADATA = "https://api.tibiadata.com/v2/world/%s.json"
+
+
+
